pyproject.toml
[project]
name = "image-classifier"
version = "0.1.0"
description = "Webcam capture and image saving utilities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "httpx>=0.27",
    "opencv-python>=4.10",
    "Pillow>=10.0",
    "protobuf>=5.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.9",
    "redis>=5.0",
    "sentencepiece>=0.2",
    "torch>=2.0",
    "tqdm>=4.66",
    "transformers>=4.40",
    "uvicorn[standard]>=0.30",
]

[project.scripts]
image-classifier = "image_classifier._cli:main"

[dependency-groups]
dev = [
    "fakeredis>=2.21",
    "pytest>=8.0",
    "pytest-cov>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
